Reduce public API surface by converting unused public items to internal
Analyzed MistDemo usage and changed access control to internal for items not
used by the example application. This significantly reduces the public API
surface while preserving all functionality needed by external users.

Key changes:
- Authentication utilities made internal (TokenValidation, CharacterMapEncoder, etc.)
- Configuration types made internal (MistKitConfiguration, EnhancedMistKitConfiguration)
- Infrastructure types made internal (MistKitClient, various protocols)
- Added missing EnvironmentConfig.swift for environment variable utilities
- Kept public APIs used by MistDemo (CloudKitService, token managers, model types)

+++ Sources/MistKit/CustomFieldValue.CustomFieldValuePayload.swift
@@ -59,7 +59,7 @@ extension CustomFieldValue.CustomFieldValuePayload {
     if let value = try? container.decode(String.self) {
       return .stringValue(value)
     }
-    if let value = try? container.decode(Int64.self) {
+    if let value = try? container.decode(Int.self) {
       return .int64Value(value)
     }
     if let value = try? container.decode(Double.self) {
